+++ layout/columns2.php
@@ -105,14 +105,11 @@ if (get_config('theme_boost_campus', 'showsettingsincourse') == 'yes') {
     $templatecontext['activitynode'] = theme_boost_campus_get_incourse_activity_settings();
 }
 
-// Render colums2.mustache from boost_campus.
-echo $OUTPUT->render_from_template('theme_boost_campus/columns2', $templatecontext);
-// MODIFICATION END.
-/* ORIGINAL START.
-echo $OUTPUT->render_from_template('theme_boost/columns2', $templatecontext);
-ORIGINAL END. */
+// MODIFICATION START: Handle additional layout elements.
+// This is needed to get the additional files template context without adding them to this file directly.
+ob_start();
 
-// MODIFICATION START: Require additional layout files.
+// Require additional layout files.
 // Add footer blocks and standard footer.
 require_once(__DIR__ . '/includes/footer.php');
 // Get imageareaitems config.
@@ -127,4 +124,20 @@ if (!empty($footnote)) {
     // Add footnote layout file.
     require_once(__DIR__ . '/includes/footnote.php');
 }
+
+// Get output buffer.
+$additionallayouts = ob_get_clean();
+
+// Check if there's something in the buffer.
+if ($additionallayouts == false) {
+    $additionallayouts = '';
+}
+// Add the template context.
+$templatecontext['additionallayouts'] = $additionallayouts;
+
+// Render columns2.mustache from boost_campus.
+echo $OUTPUT->render_from_template('theme_boost_campus/columns2', $templatecontext);
 // MODIFICATION END.
+/* ORIGINAL START.
+echo $OUTPUT->render_from_template('theme_boost/columns2', $templatecontext);
+ORIGINAL END. */